Initial upload
Added a page template with a navbar that will be copied into every created page. This still needs so work on the css. There is currently some inconsistency with a active link changing colour if it is a drop down link. Also I havent yet glued the navbar to the top of the page